+++ client/src/widgets/chatting/Message.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import { Message } from './Message';
+import { Message } from './ui/Message';
 
 const meta = {
   component: Message,
